+++ tests/test_plugin.py
@@ -1,10 +1,9 @@
 from __future__ import annotations
 
-from typing import cast
+from typing import Callable, TypeVar
 
 import pytest
 
-from pyflp.mixer import Insert
 from pyflp.plugin import (
     AnyPlugin,
     FruityBalance,
@@ -13,6 +12,7 @@ from pyflp.plugin import (
     FruitySend,
     FruitySoftClipper,
     FruityStereoEnhancer,
+    PluginID,
     Soundgoodizer,
     SoundgoodizerMode,
     StereoEnhancerEffectPosition,
@@ -20,32 +20,29 @@ from pyflp.plugin import (
     VSTPlugin,
 )
 
+from .conftest import ModelFixture
 
-@pytest.fixture(scope="session")
-def plugins(inserts: tuple[Insert, ...]):
-    return [slot.plugin for slot in inserts[19]][:7]
-
-
-def test_plugin_types(plugins: tuple[AnyPlugin, ...]):
-    assert [type(plugin) for plugin in plugins] == [
-        FruityBalance,
-        FruityFastDist,
-        FruitySend,
-        FruitySoftClipper,
-        FruityStereoEnhancer,
-        Soundgoodizer,
-        VSTPlugin,
-    ]
+T = TypeVar("T", bound=AnyPlugin)
+PluginFixture = Callable[[str, type[T]], T]
 
 
-def test_fruity_balance(plugins: tuple[AnyPlugin, ...]):
-    fruity_balance = cast(FruityBalance, plugins[0])
+@pytest.fixture
+def plugin(get_model: ModelFixture):
+    def wrapper(preset_file: str, type: type[T]):
+        suffix = f"plugins/{preset_file}"
+        return get_model(suffix, type, PluginID.Data, PluginID.Wrapper)
+
+    return wrapper
+
+
+def test_fruity_balance(plugin: PluginFixture[FruityBalance]):
+    fruity_balance = plugin("fruity-balance.fst", FruityBalance)
     assert fruity_balance.volume == 256
     assert fruity_balance.pan == 0
 
 
-def test_fruity_fast_dist(plugins: tuple[AnyPlugin, ...]):
-    fruity_fast_dist = cast(FruityFastDist, plugins[1])
+def test_fruity_fast_dist(plugin: PluginFixture[FruityFastDist]):
+    fruity_fast_dist = plugin("fruity-fast-dist.fst", FruityFastDist)
     assert fruity_fast_dist.pre == 128
     assert fruity_fast_dist.threshold == 10
     assert fruity_fast_dist.kind == FruityFastDistKind.A
@@ -53,22 +50,22 @@ def test_fruity_fast_dist(plugins: tuple[AnyPlugin, ...]):
     assert fruity_fast_dist.post == 128
 
 
-def test_fruity_send(plugins: tuple[AnyPlugin, ...]):
-    fruity_send = cast(FruitySend, plugins[2])
+def test_fruity_send(plugin: PluginFixture[FruitySend]):
+    fruity_send = plugin("fruity-send.fst", FruitySend)
     assert fruity_send.dry == 256
     assert fruity_send.send_to == -1
     assert fruity_send.pan == 0
     assert fruity_send.volume == 256
 
 
-def test_fruity_soft_clipper(plugins: tuple[AnyPlugin, ...]):
-    fruity_soft_clipper = cast(FruitySoftClipper, plugins[3])
+def test_fruity_soft_clipper(plugin: PluginFixture[FruitySoftClipper]):
+    fruity_soft_clipper = plugin("fruity-soft-clipper.fst", FruitySoftClipper)
     assert fruity_soft_clipper.threshold == 100
     assert fruity_soft_clipper.post == 128
 
 
-def test_fruity_stereo_enhancer(plugins: tuple[AnyPlugin, ...]):
-    fruity_stereo_enhancer = cast(FruityStereoEnhancer, plugins[4])
+def test_fruity_stereo_enhancer(plugin: PluginFixture[FruityStereoEnhancer]):
+    fruity_stereo_enhancer = plugin("fruity-stereo-enhancer.fst", FruityStereoEnhancer)
     assert fruity_stereo_enhancer.stereo_separation == 0
     assert fruity_stereo_enhancer.effect_position == StereoEnhancerEffectPosition.Post
     assert fruity_stereo_enhancer.phase_offset == 0
@@ -77,17 +74,17 @@ def test_fruity_stereo_enhancer(plugins: tuple[AnyPlugin, ...]):
     assert fruity_stereo_enhancer.volume == 256
 
 
-def test_soundgoodizer(plugins: tuple[AnyPlugin, ...]):
-    soundgoodizer = cast(Soundgoodizer, plugins[5])
+def test_soundgoodizer(plugin: PluginFixture[Soundgoodizer]):
+    soundgoodizer = plugin("soundgoodizer.fst", Soundgoodizer)
     assert soundgoodizer.amount == 600
     assert soundgoodizer.mode == SoundgoodizerMode.A
 
 
-def test_vst_plugin(plugins: tuple[AnyPlugin, ...]):
-    ott = cast(VSTPlugin, plugins[6])
-    assert ott.name == "OTT"
-    assert ott.vendor == "Xfer Records"
-    assert ott.plugin_path == r"C:\Program Files\Common Files\VST3\OTT.vst3"
-
-    with pytest.raises(KeyError):
-        ott.fourcc
+def test_vst_plugin(plugin: PluginFixture[VSTPlugin]):
+    djmfilter = plugin("xfer-djmfilter.fst", VSTPlugin)
+    assert djmfilter.name == "DJMFilter"
+    assert djmfilter.vendor == "Xfer Records"
+    assert (
+        djmfilter.plugin_path
+        == r"C:\Program Files\Common Files\VST2\Xfer Records\DJMFilter_x64.dll"
+    )